innago-uisp-integration/requirements.txt
requests>=2.28.0
pyyaml>=6.0
schedule>=1.2.0
